Drug Prediction/Drug Prediction Model/requirements.txt
python-dotenv
matplotlib
tqdm
networkx
rdkit
rdkit-pypi
torch 
torchvision
torchaudio
torch-geometric